os: fix getConfigDir on Android
getenv("HOME") always return null on android and the homeDir variable
assertion will always abort the program. This change makes it look for
the home directory in the users password database if the directory is
not in the HOME environment variable.

Also, don't append the ".config" directory for android, the settings are
usually stored in a folder under the /data directory.