cmake: Allow to build against system libraries.
From the outset of this project I clearly stated that we bundle
dependencies on unices to prevent symbol collisions when tracing
applications that link against other versions of these libraries.

Unfortunately most Linux packagers just assume business as usual and
mutilate apitrace source tree to build against system libraries, without
caring to read the comments or understand what apitrace is about.

So make it easier to shoot themselves on the foot.  But at least shove
a fat warning on their faces.  :-P