sed: @file reads pattern or replacement from a file. Useful for shader replacement.
Replacing shaders isn't easily done on the commandline, so add "@file()" to
tell apitrace sed to read pattern or replacement from a file.
This replacement is implemented on String-type nodes, so apitrace sed now
effectively works on string elements in addition to enum elements.

The pattern/replacement separator can be a slash or any other character,
allowing use of slashes in the file path.

Signed-off-by: Arthur Huillet <ahuillet@nvidia.com>