retrace: Add an option for parallel snapshotting.
When this new -t or --snapshot-threaded option is specified, the
retracer will spawn as many threads as the available cores on the
machine, and will encode up to that number of PNG files at once.

Even on GPU-intensive traces, this patch allows a better use of the
resources by having the CPU and GPU work at the same time instead of
waiting for each other.