gltrace: Fix glBitmap / glDrawPixels(GL_BITMAP) blob size calculation.
A superficial reading of the OpenGL specification gives the idea that
GL_UNPACK_ALIGMENT should be ignored in some cases, but if one takes in
consideration all the different constraints, it actually follows that
the row stride must be aligned to unpack alignment, one way or the other.

This is particularly true for GL_BITMAPs, whose size was being
underestimated if GL_UNPACK_ALIGMENT != 1.

This also matches what Mesa does.