scripts: Fix image diff highlighting.
Differences on certain channels were not being given enough weight when
computing the alpha mask, so differences that were quite obvious to a
human were not always being highlighted.

The error was introduced by me several years ago when trying to mimic
ImageMagick compare algorithm while mapping to the operations provided
by Python PIL extension module, to avoid doing arithmetic in native
Python, which would kill performance.  But at least nowadays Python
PIL/Pillow modules have lots of primitive operations that allow to get
more accurate results without compromising performance, so it was
relatively straightforward to fix this.

Note we only mimic but never really tried to match exactly ImageMagick
compare behavior.  In particular, ImageMagic applies a step function
when thresholding, when we do a linear ramp between 0 and the threshold,
so that errors are not over the threshold get a proportional amount of
red.