gltrace: Drop support for user memory arrays with NV_vertex_program.
NV_vertex_program is vendor specific and deprecated.  Drop support for
it in order to reduce complexity and make way for better support for
modern OpenGL.

glVertexAttribPointerNV calls will still be traced, but the
corresponding fake memcpy calls will not.

Old traces still work fine.  So if necessary an older build of apitrace
can be used instead.