glretrace: Remove ARB_direct_state_access path in getTextureTarget.
If the effective type of the texture is GL_BUFFER_TEXTURE, then it is an
error to use it with glGetTextureParameteriv. This means we would have
to check for an error and assume it is a texture buffer if there was
one, in which case this isn't any cleaner than the other path.

For https://github.com/apitrace/apitrace/issues/427