retrace: support for dumping multiple snapshots (v3)
Usually if an app is using MRT, we want to dump (and diff) *all* the
render targets to track down where things are going wrong.  Also dumps
depth and stencil buffers.

When the --mrt (or -m) argument is specified, all render targets plus
depth and/or stencil are dumped, with the suffix -mrtN/-z/-s.  Otherwise
the behavior is as before, only mrt0 is dumped with no suffix.

Only implemented for GLDumper, since I don't know anything about D3D.