glws: put a loop around X window resize code
Retry the window resize if it fails.  This is an attempt to deal
with the resize error reported as "glretrace: error: drawable failed to
resize: expected 746x504, got 522x399".

After we resize the window we wait for a ConfigureNotify event.  This
_should_ indicate that the window was resized.  But in rare instances
this appears to not be the case (perhaps a subsequent ConfigureNotify
would correspond to the resize?).  This problem could depend on the
particular window manager being used.

v2: Continue using waitForEvent instead of XGetWindowAttributes -- Jose