d3dretrace: Dump constant, vertex, index and so buffers.
To be able to dump the d3d10+ buffers we have to record their original
location as a private data on newly created buffer. After that we just
have to fetch the buffers and get the private data to find out their
original address. This allows proper display of various buffers on D3D10
and D3D11. For D3D10 only constant buffers are currently dumped but only
due to laziness. D3D11 dumps constant buffers, vbs, index buffer and so
targets.