d3dretrace: Ignore non-zero DXGI adapters.
DXGI adapters are not consistent across systems, and all but the default
adapter (0) might be absent.

In particular Windows 8 onwards always enumerate the Basic Display
Driver, and dual GPUs will have more.  And when replaying the trace on
a system with fewer adapters the EnumAdapters call will fail, ultimately
leading d3dretrace to crash.

That said, at some point we should add an option to choose the adapter.