d3dretrace: Ignore incomplete SetClipPlane calls.
For a time, only one of the 4 components of the pPlane parameter were
recorded in the traces.  Such traces still are in use, and currently we
end up using unrelated/random data for the missing coefficients.

With this change we use a zero plane equation, which allows the calls to
go through and reach the drivers, but guarantees the plane will never
clip anything, even when the respective plane is enabled in
D3DRS_CLIPPLANEENABLE, so one has reproducible rendering.