gltrim: Add a new sub-command to trim a gl trace to a replayable subset
The new command can be called via "apitrace gltrim" and it will trim
a trace so that the resulting trace consists of a setup frame that contains
the execution of the calls required to recreate the state at the beginning
of the (first) target frame so that its graphical output is the same like
the output of the corresponding frame in the original trace.

If this is not successful, one can force additional frames to be added to the
setup frame.

v2: - remove tests from main repository
    - replace include protectores by pragma
    - remove commented out code
    - add hint about limitations of the tool when running it
    - fix option indention
      (all José Fonseca)
    - update the description in the markdown file.

v3: - Add handling for queries

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>