retrace: Add `--pframe-times` option
When replaying a trace, it could be useful to analyze how much time is
needed to render and present each frame. While developing a GPU driver,
for example, developers could graph the frame time series in order to
detect regressions or improvements of driver performance over time.

The `--pframe-times` option enables profiling for frame times, meaning
that replaying will output a timestamp when a frame has been presented.

Signed-off-by: Antonio Caggiano <antonio.caggiano@collabora.com>