retrace: take end of frame shapshots always from back buffer
When specifying "--snapshot=frame" it can be assumed that the user
is interested in the renderiung output that goes to the screen and
not to some framebuffer attachment that happens to be bound to be
bound to an active framebuffer and set to be the drawbuffer.

Consequently, override the frame capture logic in case the frame-end
call triggers the snapshot to get the backbuffer contents.

Closes: #699

v2: - rename endOfFrame to backBuffer (Jose)
    - add parameter to D3D replay (Jose)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>