retrace: Add an option to force using the back buffer as snapshot source
Defaulting to reading from the back buffer when snapshots are taken at
frame frequency broke some test that only use FBOs and not always
clear the back buffer properly so that garbage was read back.
Add an option that the caller can force reading from the backbuffer,
but default to the old behaviour of reading back from the currently
bound draw buffer.

Fixes: 3d7d6e16606542e3255e03d8e99335aece7e5cdf
  retrace: take end of frame shapshots always from back buffer

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>