replay: Add option to execute and check queries
By default reading back query results is skipped, this commit adds
the option to actually read back the query result and then either
ignore it (useful for performance analysis, because reading back
the result may trigger some synchronization, e.g. with the virgl
driver), or check the result against the previously recorded
one. In order for this to make sense, we have to spin on
GL_QUERY_RESULT_AVAILABLE if the expected result is one.

v2: fix query result check logic

Closes: #715

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>