gltrim: Add a sanity check for functions names and log deviations
Since we don't check the complete function name when picking the handler
we might end up using an inappropriate handler if a function is currently
not listed, but a handler is registered for another function that happens
to have the same function name prefix.

Hence, check for common suffixes of GL function names, and print out what
handler is used when the suffix of the string is not in the expected list.

Clean up the function call mapping to give less suffix warnings.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>