Suppress ANGLE logging to stdout
There's no supported way to stop ANGLE logging to stdout, but we need to
dump things via stdout. As it uses stdout directly and we use std::cout,
it's possible to change the file descriptor to go nowhere and keep
std::cout intact.

Swapping out stdout for something else is relatively simple, but doing
so while keeping std::cout going to the original location requires
duplicating the original file descriptor and replacing the internal
buffer. That requires OS-specific calls and vendor-specific STL
extensions, so I've only implemented it for Windows so far. Other parts
of this branch are already potentially Windows-specific anyway.