config: fall back to the driver values for certain strings
getConfigString returns NULL if no config was read, but the c_str() of
an empty string is not NULL, so with a config file that doesn't specify
these strings the returned values will be empty, and we end up with false
reporting. Therefore, and return NULL if the strings are empty.