trace: Prioritize IUnknown::Release calls.
An application might call Release() on one object on one thread, and
receive a newly created object with the same point on another thread,
_before_ the Release call completes.  In such cases, replaying calls by
any static order (either enter or leave) is bound to lead to spurious
use-after-free bugs.

However this can be avoided by prioritizing Release calls, so they are
always executed before creation, as the opposite can never be intended
by a race-free application.