gltrim: move gen calls and shader setup from the last frame
One major use case for this trimming is to run the final frame in a
loop to capture some performance numbers. However, if an object is
created in the last frame, and not destroyed, then these allocations
will accumulate and may lead to memory exhaustion.

Therefore, move gen calls that are not paired with delete calls and
and shader compilation and linking calls before the last frame so that
they will not be repeated when the last frame runs in a loop.

We still may se errors in loops, because calls like glTexStorage or
glBufferStorage created immutable objects that can only be set up
once, but moving these calls before the last frame is difficult,
because then we also have to replicated the bind calls, therefore
we don't take these calls into account with this optimization.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>