#850. Add a bit of excessive space in image buffer.
This is done for a rare case of buggy driver, which writes more bytes than we expect (and specs say).
I have a particular problem on reading very low-resolution textures on Windows + AMD RX 550, and the problem totally blocks retracing for me.

Note that we write a 4-byte marker immediately after the image contents and assert that it has not changed in destructor.
So if driver is buggy, we can see it in Debug build, but Release build simply proceeds.