gltrim: Add option to replace swaps in setup with glFinish
In order to not accumulate too much work in the setup frame
add the option to replace the buffers swaps that would be
happening there with glFinish to submit the work to the
graphics card. This is useful for the Mesa-CI.

To enable this behaviour specify the flag --swap-to-finish
on the command line.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>