Work around Qt6 crash involving LTO
If link time optimization is enabled in a Qt 6 build of qapitrace,
signals are not properly connected, and a crash occurs.

Somehow an interaction between the `image` library and Qt, along with
the `INTERPROCEDURAL_OPTIMIZATION` CMake setting causes this behavior.

Simply disabling link time optimization by setting the
`INTERPROCEDURAL_OPTIMIZATION` qapitrace target property to false fixes
the issue.