Revert "Prevent hangs in XCloseDisplay."
This reverts commit f9e11f904fed8bce5f3dac7b97b4d9db6ff7be22.

I expect that XCloseDisplay is just fine with Mesa as of some point in
the last 13 years.  Regardless, if there's a problem with either GLX
or EGL then we'd want to disable cleanup in the specific WS backend,
not the frontend.

With this, glws::cleanupX() gets called again, destroying our contexts
and screen in Mesa, which means that utrace events get properly
collected and flushed so that we can do performance analysis of
apitrace replays.