Add an option to convert a trace to C code
Adds the --generate-c option that specifies a directory where glretrace
will generate a C/C++ meson project that contains the API calls of the
trace. Most of thge codegen is implemented in retrace_codegen.cpp and
OpenGL specific quirks are located in glretrace_codegen.cpp.

For replaying the trace, one simply points glretrace to the .so that can
be built from the output. glretrace will try to load trace::Blob data
from the .so.data file in the same directory as the .so.

API calls that are non-trivial to replay like glx will still use the
normal trace::Call codepath.

The .so replay only supports a subset of the options for nor al replay.
Most importantly snapshots work, but only at frame granularity.