retrace: Drop the "same call no" check in takeSnapshot().
This was in place to prevent the general case takeSnapshots() saving
over the snapshot that had already been saved at frameComplete() time,
but that path doesn't call this any more.

This now means that if you're doing looping and ask for snapshots,
you'll get files for each frame in --call-nos=false mode.  This is
useful for Mesa testing, we'd like to loop the trace a few times to
make sure the output is stable -- you don't want to be doing perf
testing on a trace if the output is UB!.