Move "has takeSnapshot() already happened?" to retrace.
When looping a frame and using --call-nos=false, we want to record the
snapshot for each frame.  But takeSnapshot() was tracking whether the
last snapshot was for this frame, to avoid overwriting a
pre-swapbuffers snapshot that the window system triggered (the good
stuff) with a post-swapbuffers snapshot (probably undefined contents).